# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: User Management Api
# during dev, should point to your local machine
host: 127.0.0.1:10010
...
basePath: /
...
schemes:
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /user/{id}:
    x-swagger-router-controller: user
    get:
      operationId: get
      description: Get a user
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        "204":
          description: User not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: Update a user
      parameters:
        - name: id
          description: The id of the user to update
          type: string
          in: path
          required: true
        - name: user
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        "204":
          description: User not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteUser
      description: Delete a user
      parameters:
        - name: id
          description: The id of the user to delete
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        "204":
          description: User not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GetUserResponse:
    required:
      - id
      - email
      - forename
      - surname
      - created
    properties:
      id:
        type: string
        minLength: 40
        maxLength: 40
        description: The unique id of the user
      email: 
        type: string
        format: email
        maxLength: 255
        description: The user's e-mail
      forename:
        type: string
        minLength: 1
        maxLength: 255
        description: The user's forename
      surname:
        type: string
        minLength: 1
        maxLength: 255
        description: The user's surname
      created:
        type: string
        format: datetime
        description: The date and time the user was added
      updated:
        type: string
        format: datetime
        description: The date and time the user was updated
  User:
    type: object
    properties:
      email:
        type: string
        format: email
        maxLength: 255
        description: The user's email address
      forename:
        type: string
        minLength: 1
        maxLength: 255
        description: The user's forename
      surname:
        type: string
        minLength: 1
        maxLength: 255
        description: The user's surname
    required:
      - email
      - forename
      - surname
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
    required:
      - success
      - description